Refactor: models entity별로 분리
