+++ controllers/user.js
@@ -35,3 +35,53 @@ export const login = async (req, res) => {
       .json({ message: err.message, status: err.statusCode });
   }
 };
+
+export async function mypageHeader(req, res) {
+  try {
+    const data = await userService.getHeader(req.userId);
+    res.status(200).json({ data });
+  } catch (error) {
+    res.status(error.statusCode || 500).json({ message: error.message });
+  }
+}
+
+export const updatePassword = async (req, res) => {
+  try {
+    const { current_password, new_password, confirm_new_password } = req.body;
+    const userId = req.userId;
+    await userService.updatePassword(
+      userId,
+      current_password,
+      new_password,
+      confirm_new_password
+    );
+    res.status(200).json({ message: '비밀번호가 변경되었습니다.' });
+  } catch (error) {
+    res.status(error.statusCode || 500).json({ message: error.message });
+  }
+};
+
+export async function updateMypage(req, res) {
+  try {
+    const { name, phone_number, profile_image } = req.body;
+    const userId = req.userId;
+    await userService.updateMyPage({
+      userId,
+      name,
+      phone_number,
+      profile_image,
+    });
+    res.status(200).json({ message: '성공!' });
+  } catch (error) {
+    res.status(error.statusCode || 500).json({ message: error.message });
+  }
+}
+
+export async function myPage(req, res) {
+  try {
+    const data = await userService.getMyPage(req.userId);
+    res.status(200).json({ data });
+  } catch (error) {
+    res.status(error.statusCode || 500).json({ message: error.message });
+  }
+}
