+++ routes/mypage.js
@@ -1,16 +1,18 @@
 import express from 'express';
-import * as myPageController from '../controllers/mypage.js';
+import * as userController from '../controllers/user.js';
+import * as reservationController from '../controllers/reservation.js';
+import * as roomController from '../controllers/room.js';
 import { validateToken } from '../middleware/authorization.js';
 
 const router = express.Router();
 
 router.use(validateToken);
 
-router.get('/mypage', myPageController.myPage);
-router.put('/mypage', myPageController.updateMypage);
-router.get('/mypage/header', myPageController.mypageHeader);
-router.put('/mypage/password', myPageController.updatePassword);
-router.get('/mypage/like', myPageController.wishList);
-router.get('/mypage/bookings', myPageController.reservationList);
+router.get('/mypage', userController.myPage);
+router.put('/mypage', userController.updateMypage);
+router.get('/mypage/header', userController.mypageHeader);
+router.put('/mypage/password', userController.updatePassword);
+router.get('/mypage/like', roomController.wishList);
+router.get('/mypage/bookings', reservationController.reservationList);
 
 export default router;
