+++ services/room.js
@@ -1,4 +1,5 @@
-import * as roomRepositroy from '../models/room.js';
+import * as roomRepository from '../models/room.js';
+import * as roomsImageRepository from '../models/rooms-image.js';
 
 export async function accommodationList(userId, query) {
   const date = { start_date: query.start_date, end_date: query.end_date };
@@ -16,7 +17,7 @@ export async function accommodationList(userId, query) {
   const sortKeyword = query.sort;
 
   const [accommodationList, accommodationCount] =
-    await roomRepositroy.readAllAccommodations(
+    await roomRepository.readAllAccommodations(
       userId,
       date,
       keyword,
@@ -33,17 +34,17 @@ export async function accommodationList(userId, query) {
 }
 
 export async function accommodationById(userId, roomsId, date) {
-  const check = await roomRepositroy.checkAccommodationId(roomsId);
+  const check = await roomRepository.checkAccommodationId(roomsId);
   if (!check.length) {
     const error = new Error('해당 페이지가 존재하지 않습니다.');
     error.statusCode = 404;
     throw error;
   } else {
-    const accommodation = await roomRepositroy.readAccommodationById(
+    const accommodation = await roomRepository.readAccommodationById(
       userId,
       roomsId
     );
-    const roomList = await roomRepositroy.readRooms(date, roomsId);
+    const roomList = await roomRepository.readRooms(date, roomsId);
     if (roomList.length) {
       accommodation[0].room = roomList[0].room;
       return accommodation;
@@ -55,21 +56,21 @@ export async function accommodationById(userId, roomsId, date) {
 }
 
 export async function accommodationLike(userId, roomId) {
-  const check = await roomRepositroy.checkLike(userId, roomId);
+  const check = await roomRepository.checkLike(userId, roomId);
   if (!check) {
-    await roomRepositroy.createLike(userId, roomId);
+    await roomRepository.createLike(userId, roomId);
     return { isLike: true };
   } else {
-    await roomRepositroy.updateLike(userId, roomId, !check.isLike);
-    const isLike = await roomRepositroy.checkLike(userId, roomId);
+    await roomRepository.updateLike(userId, roomId, !check.isLike);
+    const isLike = await roomRepository.checkLike(userId, roomId);
     return { isLike: Boolean(isLike.isLike) };
   }
 }
 
 export async function roomById(id, date) {
-  const check = await roomRepositroy.roomCheck(id);
+  const check = await roomRepository.roomCheck(id);
   if (!!check.length) {
-    const result = await roomRepositroy.readRoomById(id, date);
+    const result = await roomRepository.readRoomById(id, date);
     return result;
   } else {
     const error = new Error('Page Not Found');
@@ -79,8 +80,8 @@ export async function roomById(id, date) {
 }
 
 export async function reservationInfo(roomTypeId, userId, date) {
-  const check = await roomRepositroy.roomCheck(roomTypeId);
-  const duplicateCheck = await roomRepositroy.checkReservation(
+  const check = await roomRepository.roomCheck(roomTypeId);
+  const duplicateCheck = await roomRepository.checkReservation(
     roomTypeId,
     date.start_date,
     date.end_date
@@ -91,7 +92,7 @@ export async function reservationInfo(roomTypeId, userId, date) {
       error.statusCode = 400;
       throw error;
     }
-    const data = await roomRepositroy.getRoomTypeByUserId(
+    const data = await roomRepository.getRoomTypeByUserId(
       roomTypeId,
       userId,
       date
@@ -106,7 +107,7 @@ export async function reservationInfo(roomTypeId, userId, date) {
 
 export async function payment(userId, roomTypeId, bookingInfo) {
   // 해당 날짜에 예약이 가능한지 추가 확인을 한다.
-  const check = await roomRepositroy.checkReservation(
+  const check = await roomRepository.checkReservation(
     roomTypeId,
     bookingInfo.start_date,
     bookingInfo.end_date
@@ -116,10 +117,34 @@ export async function payment(userId, roomTypeId, bookingInfo) {
     error.statusCode = 400;
     throw error;
   } else {
-    return await roomRepositroy.createResevation(
+    return await roomRepository.createResevation(
       userId,
       roomTypeId,
       bookingInfo
     );
   }
+}
+
+export async function getWishList({ userId, page, getImageAll, count }) {
+  const maxPage = await roomRepository.readWishListRowCount(userId);
+  const data = await roomRepository.readWishList(userId, page, count);
+  if (getImageAll === '1') {
+    for (let i = 0; i < data.length; i++) {
+      data[i].image = await roomsImageRepository.readAccommodationImages(
+        data[i].rooms_id
+      );
+      data[i].image = data[i].image.map(image => image.image);
+    }
+  } else {
+    for (let i = 0; i < data.length; i++) {
+      const image = await roomsImageRepository.readAccommodationImage(
+        data[i].rooms_id
+      );
+      data[i].image = image[0].image;
+    }
+  }
+  return {
+    data,
+    maxPage: Math.ceil(maxPage[0].cnt / count),
+  };
 }
