+++ services/user.js
@@ -81,3 +81,56 @@ export const login = async (email, password) => {
   const token = jwt.sign({ id: userInfo.id }, process.env.SECRET_KEY);
   return token;
 };
+
+export async function getMyPage(userId) {
+  const userInfo = await userRepository.readMyPage(userId);
+  return userInfo;
+}
+
+export async function updateMyPage(userInfo) {
+  const phoneNumberValidation = new RegExp(
+    '^01([0|1|6|7|8|9]?)-?([0-9]{3,4})-?([0-9]{4})'
+  );
+  if (!phoneNumberValidation.test(userInfo.phone_number)) {
+    const err = new Error('PHONE_NUMBER_IS_NOT_VALID');
+    err.statusCode = 409;
+    throw err;
+  }
+  return await userRepository.updateMyPage(userInfo);
+}
+
+export const updatePassword = async (
+  userId,
+  password,
+  newPassword,
+  confirmNewPassword
+) => {
+  const userPassword = await userRepository.getUserPasswordbyId(userId);
+  const check = await bcrypt.compare(password, userPassword[0].password);
+  const salt = await bcrypt.genSalt();
+  if (!check) {
+    const error = new Error('비밀번호가 틀렸습니다.');
+    error.statusCode = 400;
+    throw error;
+  }
+  if (!(newPassword === confirmNewPassword)) {
+    const error = new Error('비밀번호가 일치하지 않습니다.');
+    error.statusCode = 400;
+    throw error;
+  }
+
+  const encryptPassword = await bcrypt.hash(newPassword, salt);
+
+  return await userRepository.updatePassword(userId, encryptPassword);
+};
+
+export async function getHeader(userId) {
+  const [check] = await userRepository.getUserbyId(userId);
+  if (!check) {
+    const error = new Error('User Not Found');
+    error.statusCode = 400;
+    throw error;
+  }
+  const result = await userRepository.readMyPage(userId);
+  return result;
+}
